+++ angelsaddons-crawlertrain/changelog.txt
@@ -1,5 +1,5 @@
 ---------------------------------------------------------------------------------------------------
-Version: 0.1.0
+Version: 0.1.1
 Date: ???
   Changes:
     - Renamed wagons to more standard Crawler train
