+++ angelsaddons-crawlertrain/data-updates.lua
@@ -1,3 +1,5 @@
+local funcs = require("prototypes/crawler-train-functions")
+
 if angelsmods.industries then
   if angelsmods.industries.components then
   -- Todo: implement proper components for the train
@@ -8,43 +10,10 @@ if angelsmods.industries then
   if angelsmods.industries.overhaul and angelsmods.industries.tech then
   -- OV.set_science_pack("angels-crawler-train", "datacore-logistic-1", 2)
   end
-  data.raw["item-subgroup"]["angels-crawler-train"].group = "angels-vehicles"
-  data.raw["item-subgroup"]["angels-crawler-train"].order = "bc[crawler-train]"
+  data.raw["item-subgroup"]["angels-vehicle-train-crawler"].group = "angels-vehicles"
+  data.raw["item-subgroup"]["angels-vehicle-train-crawler"].order = "bc[crawler-train]"
 
-  -- Update equipment grid
-  local function update_grid(grid, add, remove)
-    local function flip_t(tab)
-      local new_t = {}
-      for k, v in pairs(tab) do
-        new_t[v] = k
-      end
-      return new_t
-    end
-    if type(add) == "string" then
-      add = {add}
-    end
-    if type(remove) == "string" then
-      remove = {remove}
-    end
-
-    add = add or {}
-    remove = remove or {}
-
-    local equipgrid = data.raw["equipment-grid"][grid].equipment_categories
-    local flip_equipgrid = flip_t(equipgrid)
-    for name, value in pairs(remove) do
-      if flip_equipgrid[value] then
-        equipgrid[flip_equipgrid[value]] = nil
-      end
-    end
-    for name, value in pairs(add) do
-      if not flip_equipgrid[value] then
-        table.insert(equipgrid, value)
-      end
-    end
-  end
-
-  update_grid(
+  funcs.update_equipment_grid(
     "angels-crawler-locomotive",
     {
       "angels-energy",
@@ -53,7 +22,7 @@ if angelsmods.industries then
     },
     "angels-void"
   )
-  update_grid(
+  funcs.update_equipment_grid(
     "angels-crawler-loco-wagon",
     {
       "angels-energy",
@@ -62,42 +31,65 @@ if angelsmods.industries then
     },
     "angels-void"
   )
-  update_grid(
+  funcs.update_equipment_grid(
     "angels-crawler-wagon",
     {
       "angels-energy",
-      "angels-repair",
       "angels-heavy-defense",
-      "angels-movement"
+      "angels-movement",
+
+      "angels-repair"
     },
     "angels-void"
   )
-  update_grid(
+  funcs.update_equipment_grid(
     "angels-crawler-bot-wagon",
     {
       "angels-energy",
-      "angels-repair",
       "angels-heavy-defense",
-      "angels-construction",
-      "angels-movement"
+      "angels-movement",
+
+      "angels-repair",
+      "angels-construction"
     },
     "angels-void"
   )
 end
 
 if mods["bobvehicleequipment"] then
-  -- crawler locomotive
-  table.insert(data.raw["equipment-grid"]["angels-crawler-locomotive"].equipment_categories, "train")
-  table.insert(data.raw["equipment-grid"]["angels-crawler-locomotive"].equipment_categories, "vehicle")
-  table.insert(data.raw["equipment-grid"]["angels-crawler-locomotive"].equipment_categories, "locomotive")
-  table.insert(data.raw["equipment-grid"]["angels-crawler-loco-wagon"].equipment_categories, "train")
-  table.insert(data.raw["equipment-grid"]["angels-crawler-loco-wagon"].equipment_categories, "vehicle")
-  table.insert(data.raw["equipment-grid"]["angels-crawler-loco-wagon"].equipment_categories, "locomotive")
-  -- crawler wagon
-  table.insert(data.raw["equipment-grid"]["angels-crawler-wagon"].equipment_categories, "train")
-  table.insert(data.raw["equipment-grid"]["angels-crawler-wagon"].equipment_categories, "vehicle")
-  -- crawler bot wagon
-  table.insert(data.raw["equipment-grid"]["angels-crawler-bot-wagon"].equipment_categories, "train")
-  table.insert(data.raw["equipment-grid"]["angels-crawler-bot-wagon"].equipment_categories, "vehicle")
-  table.insert(data.raw["equipment-grid"]["angels-crawler-bot-wagon"].equipment_categories, "cargo-wagon")
+  funcs.update_equipment_grid(
+    "angels-crawler-locomotive",
+    {
+      "train",
+      "vehicle",
+      "locomotive"
+    },
+    "angels-void"
+  )
+  funcs.update_equipment_grid(
+    "angels-crawler-loco-wagon",
+    {
+      "train",
+      "vehicle",
+      "locomotive"
+    },
+    "angels-void"
+  )
+  funcs.update_equipment_grid(
+    "angels-crawler-wagon",
+    {
+      "train",
+      "vehicle"
+    },
+    "angels-void"
+  )
+  funcs.update_equipment_grid(
+    "angels-crawler-bot-wagon",
+    {
+      "train",
+      "vehicle",
+      "cargo-wagon"
+    },
+    "angels-void"
+  )
 end
