+++ angelsaddons-crawlertrain/data.lua
@@ -8,10 +8,14 @@ angelsmods.addons.crawlertrain.number_tint = {r = 0.95, g = 0.65, b = 0.25, a = 
 tech_unlocks = {}
 
 require("prototypes.category")
-require("prototypes.equipment-grid")
-require("prototypes.entities.crawler-train")
+
+require("prototypes.entities.crawler-locomotive")
+require("prototypes.entities.crawler-locomotive-wagon")
 require("prototypes.entities.crawler-wagon")
+require("prototypes.entities.crawler-bot-wagon")
+
 require("prototypes.recipes.crawler-recipe")
+
 require("prototypes.technology.crawler-technology")
 
 tech_unlocks = nil
