+++ angelsaddons-crawlertrain/info.json
@@ -8,7 +8,6 @@
   "homepage": "",
   "description": "Adds a Crawler themed train",
   "dependencies": [
-    "base >= 0.18.0",
-    "? WagonColors"
+    "base >= 0.18.18"
   ]
 }
