+++ angelsaddons-crawlertrain/prototypes/category.lua
@@ -11,17 +11,17 @@ end
 
 data:extend(
   {
-    {
-      type = "item-subgroup",
-      name = "angels-crawler-train",
-      group = mods.angelsindustries and "angels-vehicles" or "logistics",
-      order = "fa"
-    },
+    --{
+    --  type = "item-subgroup",
+    --  name = "angels-crawler-train",
+    --  group = mods.angelsindustries and "angels-vehicles" or "logistics",
+    --  order = "fa"
+    --},
     {
       type = "item-subgroup",
       name = "angels-vehicle-train-crawler",
       group = mods.angelsindustries and "angels-vehicles" or "logistics",
-      order = "bb[crawler-train]" -- order bc reserved for smeltingtrain (see addon data-update)
+      order = mods.angelsindustries and "bb[crawler-train]" or "z[angel-train]-a[crawler]" -- order bc reserved for smeltingtrain (see addon data-update)
     } -- order bd reserved for petrotrain (see addon data-update)
   }
 )
