--- angelsaddons-crawlertrain/prototypes/equipment-grid.lua
@@ -1,32 +1,0 @@
-data:extend(
-  {
-    {
-      type = "equipment-grid",
-      name = "angels-crawler-locomotive",
-      width = 6,
-      height = 6,
-      equipment_categories = {"angels-void"}
-    },
-    {
-      type = "equipment-grid",
-      name = "angels-crawler-loco-wagon",
-      width = 6,
-      height = 6,
-      equipment_categories = {"angels-void"}
-    },
-    {
-      type = "equipment-grid",
-      name = "angels-crawler-wagon",
-      width = 12,
-      height = 12,
-      equipment_categories = {"angels-void"}
-    },
-    {
-      type = "equipment-grid",
-      name = "angels-crawler-bot-wagon",
-      width = 12,
-      height = 12,
-      equipment_categories = {"angels-void"}
-    }
-  }
-)
