+++ angelsaddons-petrotrain/info.json
@@ -8,6 +8,6 @@
   "homepage": "",
   "description": "Adds a Petrochem themed train",
   "dependencies": [
-    "base >= 0.18.0"
+    "base >= 0.18.18"
   ]
 }
