+++ angelsaddons-petrotrain/prototypes/petro-category.lua
@@ -1,11 +1,11 @@
 if not data.raw["equipment-category"]["angels-void"] then
-	data:extend(
-	{
-	  {
-		type = "equipment-category",
-		name = "angels-void",
-	  },
-	})
+  data:extend(
+  {
+    {
+      type = "equipment-category",
+      name = "angels-void",
+    },
+  })
 end
 
 data:extend(
@@ -13,7 +13,7 @@ data:extend(
   {
     type = "item-subgroup",
     name = "angels-petrotrain",
-	group = "logistics",
-	order = "za",
+    group = "logistics",
+    order = "z[angel-train]-b[petro]",
   },
 })
