+++ angelsaddons-smeltingtrain/data-updates.lua
@@ -1,17 +1,68 @@
+local funcs = require("prototypes/smelting-train-functions")
+
 if angelsmods.industries then
   if angelsmods.industries.components then
     require("prototypes.recipes.smeltingtrain-updates")
   end
   data.raw["item-subgroup"]["angels-smeltingtrain"].group = "angels-vehicles"
   data.raw["item-subgroup"]["angels-smeltingtrain"].order = "bc[smelting-train]"
+
+  funcs.update_equipment_grid(
+    "angels-smelting-locomotive",
+    {
+      "angels-energy",
+      "angels-heavy-defense",
+      "angels-movement"
+    },
+    "angels-void"
+  )
+  funcs.update_equipment_grid(
+    "angels-smelting-locomotive-tender",
+    {
+      "angels-energy",
+      "angels-heavy-defense",
+      "angels-movement"
+    },
+    "angels-void"
+  )
+  funcs.update_equipment_grid(
+    "angels-smelting-wagon",
+    {
+      "angels-energy",
+      "angels-heavy-defense",
+      "angels-movement",
+
+      "angels-repair"
+    },
+    "angels-void"
+  )
 end
 
 if mods['bobvehicleequipment'] then
-  -- smelting locomotive
-  table.insert(data.raw["equipment-grid"]["angels-smelting-locomotive"].equipment_categories,"train")
-  table.insert(data.raw["equipment-grid"]["angels-smelting-locomotive"].equipment_categories,"vehicle")
-  table.insert(data.raw["equipment-grid"]["angels-smelting-locomotive"].equipment_categories,"locomotive")
-  -- smelting wagon
-  table.insert(data.raw["equipment-grid"]["angels-smelting-wagon"].equipment_categories,"train")
-  table.insert(data.raw["equipment-grid"]["angels-smelting-wagon"].equipment_categories,"vehicle")
+  funcs.update_equipment_grid(
+    "angels-smelting-locomotive",
+    {
+      "train",
+      "vehicle",
+      "locomotive"
+    },
+    "angels-void"
+  )
+  funcs.update_equipment_grid(
+    "angels-smelting-locomotive-tender",
+    {
+      "train",
+      "vehicle",
+      "locomotive"
+    },
+    "angels-void"
+  )
+  funcs.update_equipment_grid(
+    "angels-smelting-wagon",
+    {
+      "train",
+      "vehicle",
+    },
+    "angels-void"
+  )
 end
