+++ angelsaddons-smeltingtrain/data.lua
@@ -1,5 +1,3 @@
-local __DebugAdapter = mods["debugadapter"] and require("__debugadapter__/debugadapter.lua")
-
 angelsmods = angelsmods or {}
 angelsmods.addons = angelsmods.addons or {}
 angelsmods.addons.smeltingtrain = angelsmods.addons.smeltingtrain or {}
@@ -10,8 +8,13 @@ angelsmods.addons.smeltingtrain.number_tint = {r = 1, g = 0.38, b = 0, a = 1}
 tech_unlocks = {}
 
 require("prototypes.smelting-category")
-require("prototypes.entities.smelting-train")
+
+require("prototypes.entities.smelting-locomotive")
+require("prototypes.entities.smelting-locomotive-tender")
+require("prototypes.entities.smelting-wagon")
+
 require("prototypes.recipes.smelting-recipe")
+
 require("prototypes.technology.smelting-technology")
 
 tech_unlocks = nil
